backend/requirements.txt
alembic==1.14.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.23.2
psycopg[binary]==3.2.3
python-dotenv==1.0.1
